NPC colour, gender, level manager, pcg start, new building actor types.
Colour variation in NPCS and buildings. Movement to different locations based on randomize value and tags. New location types.
Randomized male and female NPC types. Level Manager for types of actors to allow NPCs to interact with specific places

Fixing: Get the level manager to actually work and allow that to help with Daily Routine Behaviour for locations.

 Addition of PCG plugin - will use for town layout, and perhaps NPC creation.