Continued on spline layout - adjust absolute rotation, get data for building rotation
Attempting to get buildings to face the road. Need to determine front of each building and use that as the direction to orient perpendicular to the tangent of the spline and the nearest point. This is only for the front row of buildings - all others behind should follow the orientation of the building in front of it.